Convert startTime and duration of mark/measure to double, to align with user timing spec and navigation timing 2 spec.

+++ src/main/java/com/soulgalore/web/browsertime/datacollector/TimingDataCollector.java
@@ -58,7 +58,7 @@ public class TimingDataCollector {
             TimingMark end = timingRun.getMark(endMarkName);
 
             if (start != null && end != null) {
-                long duration = end.getStartTime() - start.getStartTime();
+                double duration = end.getStartTime() - start.getStartTime();
                 TimingMeasurement m = new TimingMeasurement(measurementName, start.getStartTime(), duration);
                 timingRun.addMeasurement(m);
             }
